# Add `JAXEpochIterator` and speed up `get_numpy_iterator`