--- a/keras/backend/torch/trainer.py
+++ b/keras/backend/torch/trainer.py
@@ -1,3 +1,5 @@
+import collections
+import itertools
 import warnings
 
 import numpy as np
@@ -8,8 +10,8 @@ from packaging.version import parse
 from keras import backend
 from keras import callbacks as callbacks_module
 from keras import optimizers as optimizers_module
+from keras.backend.torch.core import get_device
 from keras.trainers import data_adapters
-from keras.trainers import epoch_iterator
 from keras.trainers import trainer as base_trainer
 from keras.trainers.data_adapters import data_adapter_utils
 from keras.trainers.epoch_iterator import EpochIterator
@@ -263,9 +265,9 @@ class TorchTrainer(base_trainer.Trainer):
 
             # Run validation.
             if validation_data and self._should_eval(epoch, validation_freq):
-                # Create EpochIterator for evaluation and cache it.
+                # Create TorchEpochIterator for evaluation and cache it.
                 if getattr(self, "_eval_epoch_iterator", None) is None:
-                    self._eval_epoch_iterator = EpochIterator(
+                    self._eval_epoch_iterator = TorchEpochIterator(
                         x=val_x,
                         y=val_y,
                         sample_weight=val_sample_weight,
@@ -328,7 +330,7 @@ class TorchTrainer(base_trainer.Trainer):
             epoch_iterator = self._eval_epoch_iterator
         else:
             # Create an iterator that yields batches of input/target data.
-            epoch_iterator = EpochIterator(
+            epoch_iterator = TorchEpochIterator(
                 x=x,
                 y=y,
                 sample_weight=sample_weight,
@@ -494,11 +496,49 @@ class TorchTrainer(base_trainer.Trainer):
         return batch_outputs
 
 
-class TorchEpochIterator(epoch_iterator.EpochIterator):
+class TorchEpochIterator(EpochIterator):
     def _get_iterator(self, return_type="auto"):
         if return_type == "auto" and isinstance(
             self.data_adapter, data_adapters.TorchDataLoaderAdapter
         ):
             return self.data_adapter.get_torch_dataloader()
+        else:
+            # defaults to `np` if not `TorchDataLoaderAdapter`
+            return_type = "np"
 
+        if return_type == "np":
+            return self._prefetch_numpy_iterator(
+                super()._get_iterator(return_type)
+            )
         return super()._get_iterator(return_type)
+
+    def _prefetch_numpy_data(self, data):
+        def to_device(d):
+            return torch.as_tensor(d, device=get_device())
+
+        return tree.map_structure(to_device, data)
+
+    def _prefetch_numpy_iterator(self, numpy_iterator):
+        """Prefetch batches on device.
+
+        The idea has been borrowed from
+        `torchtnt.utils.data.CudaDataPrefetcher`
+
+        This utility takes an iterator and returns a new iterator which fills an
+        on device prefetch buffer. Eager prefetching can improve the performance
+        of training loops significantly by overlapping compute and data
+        transfer.
+        """
+        queue = collections.deque()
+
+        # If you're training on GPUs, 2 is generally the best choice because
+        # this guarantees that you can overlap a training step on GPU with a
+        # data prefetch step on CPU.
+        def enqueue(n=2):
+            for data in itertools.islice(numpy_iterator, n):
+                queue.append(self._prefetch_numpy_data(data))
+
+        enqueue(n=2)  # TODO: should we make `n` configurable?
+        while queue:
+            yield queue.popleft()
+            enqueue(1)