@james77777778 2023-12-26
> This is great! Thank you for proposing this. Maybe we could something similar for torch as well?

Sure. I have added a similar prefetching method in `TorchEpochIterator` and replaced `EpochIterator` with it.
You can check the performance gain:
https://github.com/keras-team/keras/pull/18991#issue-2055704197

Note that: the above prefetching only takes effect when it is not a `TorchDataLoaderAdapter`, as Dataloader should have its own prefetching inside.